Added a Startup page that waits for the ghost-word-server to come online before allowing the game to start.
Since the ghost-word-server is hosted on a free Heroku instance, it'll automatically go offline after a period of inactivity. The first time the page is visited after this period of inactivity, it can take a long time for Heroku to spin our server back up, so we wait on the Startup page until we receive a ping response from the server.

- Made the 'call bullshit on <player>" button wider to avoid cutting off the players' name